+++ RELEASE-NOTES.md
@@ -1,3 +1,4 @@
 ## 0.1
 
 * #1 Basic implementation
+* #3 Generic mechanism for loading Realms from configuration and DI